Including error of the wrapped exception
This is not ideal, since at moment exception message can be used to show
errors to the user, which will lead to parsing errors being printed to
the screen. However at the moment the underlying error message is
important to determine the cause of the failure for debugging. Ideally
we should have specific exceptions for different failure modes with two
set of messages, one to show to the end user and another with enough
details to debug an issue.

For now, to limit the amount of changes, including the extra details
should be sufficient.