Check PFS matrix server
The Raiden client and the PFS must be in the same matrix federation,
otherwise the PFS will not see the node as online. To verify this, check
that the PFS uses on of the `available_servers`, which should contain
all servers in the matrix federation.

The typical symptom of the error we avoid is `PFS: Source not online.
Approximate block at the time of the request ... (PFS error code:
2201)`. We had this error twice due to a matrix mismatch, so it's worth
the additional code for the check.