Relax ULID time_monotonic resolution
On some systems (i.e. Linux Subsystem for Windows), the monotonic time
resolution is larger than nanoseconds.

The requirement for the monotonic clock resolution is, that it needs to
be finer than consecutive calls to `time.clock_gettime_ns`, which should
still be satisfied with a `microsecond` (`1e-06`) resolution.

See also https://www.python.org/dev/peps/pep-0564/#annex-clocks-resolution-in-python

This fixes #6224