Use `command -v` instead of `type`
I got `make: type: Command not found` errors before, so I changed the
usage to `command -v` which is in the posix standard. This works without
errors for me and still correctly checks for the command existence.

I'm not really sure why I did get the error in the first place, since
both `bash` and `dash` support the `type` command.