Stop using `use_kwargs`
I could not get it working with recent webargs versions and it was
intransparent what it did in detail. The two replacement functions are
small and straightforward.

The new version also makes it clear that the parameters in the URL path
are not being validated by the schemas. That was already the case before
this commit, but it was really hard to spot.