Faster parallel execution of split_debug_logs.sh
I noticed that this could take minutes on CI just for the test
`test_assumption_search_user_directory_returns_federated_users`, which
is unreasonable. Timing different approaches on my machine revealed that
the semaphore approach seems to have a high overhead and is only usable
for longer running processes. So I switching to the classical way to
call GNU parallel, which is also a bit simpler IMO.

Timing from my machine:

parallel with semaphore: 120s
xargs: 40s
parallel classic: 20s

This could still be much faster, but it's ok for now.