Add claims path cli option and change claims format to jsonl
There is now a new `--claims-file-path` cli option to specify the location
of a claims file.

Also the claims file is now in jsonl (i.e. one json object per line) 
format. 
The first line contains the as of yet unused operator info.
All lines after that are claims.

Parsing of the claims has been partially converted to a generator style
to avoid having to keep the entire claims in memory multiple times.