Only warn if eth client is not supported
Other versions might work, we just did not check for compatibility, yet.
Use at your own risk.

This is suboptimal, but since we can't tell which versions are
compatible before we test them, this looks like the best compromise
on the scale between preventing from running on any untested versions
and staying silent.