Split known servers
Depends-on: raiden-network/raiden-service-bundle#186
Related: #6443, #6212

This splits the known servers list into an 'active' and 'all' part.

The 'active' part is read by the Raiden client and used to select a
server to use.

The 'all' part is used by the RSB as the federation whitelist.

The purpose of this is to allow new servers to be added without making
the Raiden client immediately try to use them while they are not yet
ready.

(cherry picked from commit 920f7822786b3b96daa3852d6414b8ae851309eb)