Stop sync thread before address manager
This fixes the flakyness of `test_payment_statuses_are_restored`. What
happened before was the following:

* The address manager was stopped and reset it's state when stopping
* The client's sync thread was still running for a short time and wrote
to the address manager.
* After the node restart, the address manager's state was unclean, which
led to misbehaving health checks
* The Presence updates were not correctly handled, since the node was
not whitelisted
* Some messages get stuck in the queue because the other node seemed
offline

While this fixes the problem, I think it is unwise to have such state
reset code as we currently have in the address manager. Removing the
objects on stop and creating them again on restart is far safer and
probably even less code.