Avoid six version conflict in scenario player
Synapse uses treq, which uses six. Unfortunately, the recent treq
version does requires six>=1.13.0 but does not declare that in the
dependencies. This works ok for the current Raiden setup, since the
versions of all dependencies are frozen and the treq version is old
enough to not require the recent six version.
However, this is not true for the SP, which only versions direct
dependencies. I added the workaround to this repo rather than in the SP
repo, so that we won't get the same breakage again when one of the
dependencies is updated here.

See also https://github.com/twisted/treq/commit/934e127a2b915bf02be86d23f2cf8a65bcdb2533#r42782781