Fix inconsistent TokenNetworkRegistryState
After deserialization, the TokenNetworkStates in `token_network_list`
and and `tokennetworkaddresses_to_tokennetworks` were equal but not
identical.  When one of them was mutated, the other stayed unchanged,
leading to inconsistent results for the same token network.

To avoid this problem, only the `token_network_list` attribute is
serialized and `tokennetworkaddresses_to_tokennetworks` is derived from
that after deserialization.

It is still possible to modify these two attributes separately, but by
adding and using the `add_token_network` method, this is made less
likely. I did not find a way to completely remove the possibility
inconsistencies without breaking our (de)serialization mechanism.