Fix bf1 request_count mismatch
This fixes #6590

Since removal of local routing, the amount of PFS requests has changed,
because finding out that a transfer is not possible now needs a
Routing Request to the PFS in all cases.

This commit also contains a number of yaml fixes -- pre-commit
validation wouldn't accept jinja templating fields (`{{ ... }}`)
without quotes.