Check for suitable channel before asking PFS
When there is no direct channel to the target, the node can't know if it
can reach the target without asking the PFS. But in some cases, it can
know that there can't be a route just by looking at it's direct
channels. Skip the PFS request in this case.

Closes https://github.com/raiden-network/light-client/issues/2239.