Adjust exp backoff according to max request time
For each block range, we make multiple requests with different filters.
Before this commit, only the last request duration has been used to
adjust the batch size. But the point of the adjustment is to prevent any
request from running into a timeout, so adjusting based on the maximal
duration of all requests is a better way to achieve this.