Work around flaky eth node startup in CI
We always had problems with eth nodes that started without errors but
never reliably processed blocks. This is the reason why we allow test
retries when they fail during startup. The recent forced upgrade to
openethereum made this problem more pronounced, so that a high
percentage of CI runs failed due to a hanging node startup.

The combination of the following changes in this PR seems to be
effective:
* Allow tests to run up to 20mins (we have a separate timeout for times
without output)
* Allow up to three runs for test that fail during startup
* Reduce the allowed time for startup to 4 minutes, so that we don't
wait unnecessarily long for hanging eth clients.