Return tx_hash for `mint_for` APIs in all layers
The documentation of the public REST-API falsely claimed that the
transaction_hash will be returned in the response-body
for the endpoint at `POST /api/v1/_testing/tokens/<addr>/mint`.

Since no methods in the call stack of that endpoint returned
the transaction_hash, this behavior is implemented now -
the conversion from HexBytes to a HexString is done in the
most outer REST layer.

Resolves: #6649