Test `CustomToken` blockchain proxy
Before, the `CustomToken` proxy was not tested explicitly.
In this commit, a separate test for the `CustomToken` proxy is added,
testing very basic `mint` and `mint_for` functionality.

As of 8d2419c807dbcafd96cb6ad22b389f20d6cfb789 the CustomToken.mint_for
method returns the transaction_hash of a successful mintFor transaction,
this is tested with a simple encoding check for the return value.