Fix query balances in UDC-proxy
The `total_deposit` and `balances` mappings from the UDC
contract were assumed to represent the same in the proxy wrapper
methods. Compared to the monotonically increasing `total_deposit`,
the `balances` can decrease upon withdraws and transfers.
This was not noticed in the tests, since balance and total deposit
are the same before any withdraw or transfer is conducted.
Depositing after a modification of the internal balances
could cause a deposit transaction to raise `deposit not increasing`
in the contract, while passing the `_deposit_preconditions( )` check.