Return tx-hash for inflight tx on UDC deposit
If a race condition causes a deposit to wait for an
inflight transaction, the hash of that inflight transaction
will be returned instead of None.

The interface of the `_deposit_inflight( )` contextmanager
has been changed, so that the async result is passed to the
with block, and the transaction hash has to be
explicitly set on the `InflightTransaction.async_result`
inside of the with block.