Use click's external name for config file options
Before, the manual step of parsing the configuration file
matched configuration file options with the  internal representation
of the click arguments by deterministically transforming the names.

Now the options are matched with click's external representation,
which is accessible from an options `.opts` attribute.
If ambiguous or unknown options are present in the config file,
an exception is raised instead of ignoring the option.

Fixes: #6690