Refactor run_raiden_service argument handling
This commit will make the min_reveal_timeout depend on
the chain-id.
Since value checks of inter-dependent options in combination with
config file  parsing is hard to do with click,
the configuration should be checked for sanity on creation of the
RaidenConfig, but before any services are started.
Therefore, the initiation of the RaidenConfig has been pulled
out of run_raiden_service, so that static configuration checking
can be separated from I/O-bound environment checking and startup.