Fix gas costs for registerSecretBatch
The previous calculation did not make any sense. That has been obscured
by taking the maximum of the calculation and the gas estimate. But the
gas estimate can be wrong when other nodes try to register the same
secret at the same time.

The new approach of calculating the gas cost could be better, but the
new assert should alert us when the values go out of sync with the
expectations of the contracts.

Closes https://github.com/raiden-network/raiden/issues/6725.