Fix MacOS builds using system Python
Before this commit, the CI was using the system Python from the
MacOS executors, which is fixed to version 3.7.0.
This would cause that not all Raiden requirements were able to be
installed for the MacOS builds, due to incompatibilities
with Python 3.7.0.

As of this commit, `pyenv` is used for all MacOS builds
to install a custom version of Python.
If the CI's `py-version parameter` is set to 3.7, this will use
Python version 3.7.9.
Additionally, it is now possible to use Python 3.8.6 for MacOS builds.
Future upgrades to other Python versions are now easier, since
arbitrary versions can be passed to `pyenv install`.

Fixes: #6705