remove assert in get_room_for_id
This assert needs to be removed because webrtc essages are to be sent and exchanged before an address is healthchecked. This leads back to the same race condition as for accepting invites for rooms.
The assert was placed to show a mistaken change in the order of healthchecking and sending messages. This is currently not a problem and it is not possible anymore to have an anchor in this place for now.