Use latest chain state for balance check
When checking if the channel balance is worth enough to justify paying a
MS, we should check the latest state of the channel, not the one before
receiving the current balance proof. Otherwise, we won't send a MR for
channel when we first cross the threshold by receiving a payment.

This has not been caught by the test for two reasons:
* access to the chain state is mocked out
* the threshold is reached by making a deposit, not by receiving a
  payment

I didn't find a simple way to write a regression test, so I'll save the
time of other tasks. The code path is mostly tested and this is not too
likely to regress, so this is somewhat acceptable. Still, if you see a
nice way to write the test, please do so!

Closes https://github.com/raiden-network/raiden/issues/6736