Add `--version` as top-level CLI flag
This will print the Raiden short version when the CLI is invoked as
`raiden --version`, in the same manner `raiden version --short` will.
Rationale behind this is that printing the version caused by a  flag
instead of a subcommand is a common pattern that is expected by most
users and developers.

Fixes #6669