Refactor config-file parsing
The parsing of the config-file options was done in a way
that was not very modular and reusable.

This commit allows for introspection of what part of the
configuration parsing set a specific parameter, via the
`click.core.ParameterSource` enum that was added in click version 8.

A newly added `Parser` abstract-class allows for the construction
of options that trigger additional parsing of the CLI parameters,
for example a config file that overwrites default values, but
is overwritten by values provided by the CLI.
This behavior is implemented in the `ConfigParser` class.

The `option( )` decorator now has an additional arguments
`option_parser_cls` and `option_parser_priority`, so that an additional
`Parser` can be attached to an option as described above.

The extension of the `click` module classes is changed to
`click==8.0.0a1` release-candidate behavior,
assuming there will be no big API changes in the `8.0.0` release.