Add unit-tests for CLI parsing
Before, we did not have tests to check the parsing of CLI commands,
arguments and options.
Also, the interplay configuration file options and command-line
options was not tested.

In this commit, the added tests will check the correct prioritization
of option defaults, cli-options and config file options.

Additionally, the CliRunner fixture is changed from using
a closure to set fixed environment variables to using
it's own constructor argument.