Change smoketest option parsing
The smoketest now uses a closure and the CLI-option decorators
to utilize the complete CLI-option parsing including filling in the
proper CLI default values.

The `password_file` value in RaidenTestSetup.args now
returns a path instead of a wrapped click.File() class.
This is a compatibility-breaking change for the scenario-player.