Fix incorrectly passed value for missing default config file
The config-file option was intended to be passed as `None`
to a wrapped command function from the CLI parsing,
if it was missing and not explicitly set by the parser.
This caused an incorrect logging output and incorrect reasoning
about whether a config file was used for options or not.

This commit correctly fixes this behavior.