Add warning for `enable_monitoring` option on mainnet
A user could lose funds when monitoring is disabled
on mainnet.
Since the monitoring should not be active per default due
to higher costs, at least show a warning when the user
did not enable the monitoring while connecting to the mainnet.
The warning is not shown when the user disabled monitoring
explicitly via the CLI or a config-file.

Fixes #6779