Change services-broadcast from rooms to `to-device` messages
Before, some messages to the services (requests for monitoring,
updating of path information) were broadcasted to specific matrix
rooms via room-messages.

Now, the implementation of the service broadcast is internally changed
to be done via multicast `to-device` messages to all service addresses
that are fetched from the blockchain.

The `device-id` of the multicast message specifies the topic of the
multicast, and is similar to the former different broadcast rooms.

Resolves #6814