Add optional AddressMetadata processing
This is defining the interface of how the transport will receive
additional address-metadata that can be bound to individual messages.
This information will be used to specify more specific information
about transport specific user-ids and capabilities - that can
be forwarded to the node's transport from other peer's forwarded
metadata or from the PFS.
The transport will be able to send messages without the metadata.