Add optional AddressMetadata support in MatrixTransport
The MatrixTransport is now able to handle additional metadata
for messages that will primarily contain specific matrix user-ids
to send message to via to-device.
If a message is passed to the transport without metadata,
to-device messages will be sent to all possible user-ids on all
matrix home-servers known to the transport.

WebRTC is being disabled in this commit (temporarily).