Store RouteStates in LockedTransferSignedState instead of addresses
This will refactor the states so that the LockedTransferSignedState
stores a list of RouteStates instead of address-only paths.
This will make it easier to store and access AddressMetadata that
is contained in a received LockedTransfer on the receiver side,
as some additional metadata arguments throughout the state-machine
functions and states become void.