Fix flaky `test_matrix_transport_handles_metadata`
The test was using an additional transport in order to derive a
valid user-id that didn't match the queue-identifiers `recipient`.
The test had a bug, where the `Delivered` messages of the transport2
were not considered in the asserts.
This is now fixed by simply removing the transport2 and it's user-id
from the message sends.