Fix invoking pytest without path
Until now pytest needed to be invoked with the path to the tests
(`raiden/tests`) in order for the top level `conftest.py` file to be
read early enough for our custom CLI options to be enabled.
Also certain other hooks (e.g. temp path shortening on macOS) are
configured there and only take effect when this file is read early
enough.

Having to specify the path has been a long standing irritation for both
old and new contributors.

This moves the top level conftest file to the project root so that it's
read early enough in all cases.