Add pre-commit config version check helper
We often have the problem that the tool versions specified in the
pre-commit config drift out of sync with the general project
requirements.

Especially in the case of `mypy`, `isort` and `black` this can lead to
annoying "edit wars" between the differing versions.

This adds a utility that checks that the versions match the requirements
and adds it to the `lint` make target (thereby ensuring out of sync
versions will be caught by CI).