Remove fallback send to-device to all possible UserIDs
For invalid or non-existent AddressMetadata in the `_send_raw`
method of the transport, a fallback was used where the message
was sent out to all possible user-ids on all known Matrix servers
in the federation. This was considered a possible security
vulnerability, since a malicious matrix server wouldn't have
to follow signature verification and could be eavesdropping.
The fallback feature is therefore removed, so that messages without
address-metadata will remain the retry-queue as long as there is no
WebRTC connection established.

This commit will also fix the flaky
`test_matrix_transport_handles_metadata`, where the `Delivered`
messages of the transport2 were not considered in the asserts.

Fixes: #6953
Fixes: #6976