Fix ignoring of `recipient_metadata` for messages recovered from WAL
When messages were recovered from the WAL and enqueued in the transport
because they were not sent yet, we didn't consider the
`recipient_metadata` of the recovered `SendMessageEvent`.
Before, this was not noticed because the transport used the
fallback UserID generation scheme transparently.
This mechanism is now removed, which made the tests in
`raiden/tests/integration/test_send_queued_messages.py` fail.