Explicitly set test root dir in scenario player tests
Since #6991 a `conftest.py` file exists in the Raiden project root.
This seems to cause the scenario player tests to pick that up as the
test root dir (since the SP repo is cloned into the raiden checkout) and
use the raiden conftest file.
This produces unexpected results (infinite recursion).

Explicity passing a rootdir fixes the issue.