Validate address metadata in RouteStates
Until now `RouteState` had no validation on the `address_to_metadata`
field (unlike `RouteMetadata` which is used in the transport).
In cases where invalid metadata were present this would lead to stuck
transfers that would need to wait for lock expiration.

Adding the validation now allows either to select a route with valid
metadata for all hops or immediate user feedback that the transfer can't
be performed.

See #7049 for more details.