Fix address-metadata state mutation in message verification
States should never be mutated when the mutation is not the cause
of a state-change.
The RouteState.address_to_metadata dictionary was passed to the
RouteMetadata message container, and the mutable object was altered
upon verification of the metadata signatures.

This could cause losing of invalid metadata in the senders state,
once the metadata was passed to other nodes.

Fixes #7061