Use canonicaljson and DictSerializer for Metadata hash generation
Before this commit, the serialization-schema for hashing of the
RouteMetadata was hard-coded, and the hashing of the Metadata
was using an inconsistent mixture of canonicaljson and rlp-encoding
for the hash-generation.

Now, the whole message is first serialized with the already employed
DictSerializer where typing information is stripped and then a canonical
serialized representation is built with canonicaljson.