Pass additional metadata through the state machine
Before, a Mediator could serialize unknown metadata fields from
incoming LockedTransfer messages, but this metadata was not stored
in the state-machine for later inclusion in following LockTransfer
messages for the next hop.

New arguments have been added throughout the state-machine to
pass the additional, unknown metadata to the corresponding
creation of SendEvents and StateTransitions.