Fix retry-queue's enqueue() blocked by semaphore
Before, the retry-queues semaphore had to be acquired in order to
enqueue messages.
This caused the thought to be asynchronous enqueue() method and
all methods calling it to be dependent on successfully processing
all messages in the queues greenlet that processes the queue.

Effectively this resulted in a synchronous enqueue() call and could
cause the state-machine to block during long-running requests or retries
to the transport layer.

The retry-queues lock was removed, since it was not necessary at all:
as long as there is only ever one retry-queue instance per channel
running, which is not the case due to the logic in `_get_retrier()`.

Fixes: #7120