Add read-only Metadata forwarding and CJSON hashing scheme
This will add a `original_data` field on the `class Metadata`
message-dataclass. This field can contain the original
metadata-dict as received from a previous node.
The serialization-scheme of the Metadata class considers the existence
of the original_data and will ignore other fields on dumping and thus
only dump the original_data.

The originally received metadata dict is now passed through the
state-machine for later inclusion in following LockTransfer
messages for the next hop.
New arguments have been added throughout the state-machine to
pass the `previous_metadata` to the corresponding
creation of SendEvents and StateTransitions.

Before this commit, the serialization-schema for hashing of the
RouteMetadata was hard-coded, and the hashing of the Metadata
was using an inconsistent mixture of canonicaljson and rlp-encoding
for the hash-generation.
Now, the whole message is first serialized with the already employed
DictSerializer where typing information is stripped and then a canonical
serialized representation is built with canonicaljson.

Fixes: #7113