Remove route-pruning for mediators
Before this commit, a mediator would modify the routes in the metadata
to prune the graph of paths before passing the metadata to the next hop
as part of the mediated locked-transfer.

Now, the route in the metadata is solely determined by the
initiator of a payment and passed on as is by all mediators.

Fixes: #7090