Remove dependency on pruned routes
Before, we were assuming pruned routes being passed to a node,
where we extracted the next hop in a route by indexed-based access
to the route-list.

Now, a node searches for the index of it's own address in a route-list
and extracts the next hop based on that.
This is a backwards compatible change, since the next hop's own
address was not being pruned by a node - that way the next hop still
finds itself in the list.