Move test from integration to unit-tests
Whether a mediator forwards metadata as it received it
in a transfer can also be tested on the state-machine level
and does not need be tested with an integration test.

Also, the integration-test was using monkeypatching that
was could not easily be adapted to recent changes to how the
metadata is passed.