Get ContractSettled amounts from blockchain event directly
Before, the ChannelSettled event did not emit the
transferred-amounts for both participants of the
settle-event from the blockchain.
As of https://github.com/raiden-network/raiden-contracts/pull/1126
the amounts are also emitted directly, and therefore do not
have to be polled from the blockchain again.
This also eliminates the need to pass the `proxy_manager` and
the `current_confirmed_head` down to the
`blockchainevent_to_state_change`. The event handler only
processes blockchain events up to the currently confirmed block,
so reorgs are only an issue when querying the blockchain specifically.

Additionally, the ChannelSettled event now contains both
participant's addresses
( https://github.com/raiden-network/raiden-contracts/pull/1494 ),
since otherwise the amounts and lockroots arguments could not be
associated with the channel-participants. This is reflected
in the decoding of the blockchain event.