Refactor waiting module
The waiting module showed a lot of code duplication that could be
abstracted.
Now there is a custom interface `ChannelStateCondition` where
custom conditions can be implemented, logically combined
("and" and "or") and handed to `ChannelStateWaiter` instances.

The existing `wait_*` functions have been re-implemented using
the new mechanism internally, but observe the same interface as before
for compatibility reasons. The functions could be factored out
at a later point.

The waiting is extensively used in the `raiden.api.python` module
as well as in tests.