Add transparent coop-settle attempt to Python-API's `close` endpoint
The off-band coop-settle protocol is an optimization for
two partner nodes to agree on the parameters for a channel-settle.
Therefore it will be transparently initiated as soon as a user
wants to close a channel. If the partner is offline or the
withdraws related to the coop-settle expire for whatever reason,
the on-chain settlement lifecycle is used.