Adapt scenarios to use `assert_channel_settled_event` for coop-settle
Before, there was no coop-settle, so that the settlement of
channels was asserted by querying for a ChannelClosed event.

Coop-settle is now the default mode for a channel closure.
A successful coop-settle life-cycle does not emit a ChannelClosed
event and will remove the channel-state from raiden.
Therefore, the only way to test for a successful coop-settle is
to poll the blockchain for the ChannelSettled event that is
emitted by the coop-settle life-cycle. Additionally, the
uncooperative settle has to be ruled out by asserting that no
ChannelClosed event was emitted on the blockchain.