Always allow WithdrawRequests with zero value
Disallowing Withdraws with total-withdraw of zero is only
necessary for regular withdraws on the initiator's side,
simply to prevent contract-calls that don't do anything.
This check is already performed before initiating a withdraw
on the initiator's side.
For the coop-settle protocol, receiving and sending of zero-valued
withdraws is necessary, so the constraint of the validation has to
be loosened. This can be safely done, since signing zero-valued
regular withdraws would not do any harm to the signer/sender of
the withdraw confirmation.