Don't ignore exceptions in WebRTCManager
The `WebRTCManager` is a `Runnable` subclass. In order for exceptions to
properly bubble up those either need to have their greenlets
(exception-) linked to the parent Runnable or need to be waited upon.

Otherwise exceptions will be silently ignored causing inconsistent
state.

This adds an exception link to the MatrixTransport.