raiden: api: initiating a coop settle can fail
If coop settle cannot be initiated (e.g. due to pending locks or
withdraws), the channel state's attribute will not be set, which is fine
and expected. We just need to take that into account.  Note that
expiration cannot have happened yet, so the only reason the
initiated_coop_settle could be None is that coop settle could not be
started at all.