raiden: allow exceptions to propagate all the way to the top
If an unhandled exception occurs (e.g. an assertion error that should
not have happened), don't try to do a clean shutdown but instead just
let the exception propagate upwards, acrosss greenlets where it will be
logged.  This will also cause the node to die, which is what we want.

Calling methods like stop() in the case of an unhandled exception is
dangerous because we might depend on resources which are not
available/functioning anymore.