raiden: transfer: be a bit flexible when computing expiration block
Other nodes may not see the same block we do, so allow for a difference
of 1 block. A mediator node can fail to use an open channel if the lock
timeout is greater than the settle timeout due to different blocks being
seen between nodes. This delays mediation for at least one block time
and therefore increases payment time.