add new 'rpc' strategy for gas price parameter
Extends the `--gas-price` parameter with a new gas strategy that queries the
connected Ethereum RPC node for the current price. This is meant to be used
especially for Arbitrum networks, as the other strategies do not work there. The
gas system works slightly different for such chains.

Furthermore there is no more a default parsing result. This means that the
command-line option still has a default, but the parsing function is strictly
evaluating the passed value.