adjust scenarios for arbitrum
Sometimes it causes issues when the Scenario Player continues right away with
the next task after opening a/some channel(s). Due to the zero confirmation
blocks this can lead to errors where a channel is not yet open. This is what we
can see in the logs of the failed scenarios.