update raiden-contracts to 0.50.0
Note that the former release candidates anticipated a version `1.0.0rc*`. This
has been changed and the final release of `raiden-contracts` has been made with
version number `0.50.0`.

Also note that in contrast to former update commits for the `arbitrum` branch,
this correctly makes usage of `requirements.in` again and compiles all
dependencies correctly.